+++ src/sessions/base-session-handler.ts
@@ -114,6 +114,12 @@ export default abstract class BaseSessionHandler {
       if (state === 'interrupted') {
         this.sdk.emit('sessionInterrupted', { sessionId: session.id, sessionType: session.sessionType, conversationId: session.conversationId });
       }
+
+      /* If the RTCPeerConnection is in a 'failed' state, it is a state of finality and we need to clean up the session. */
+      if (state === 'failed') {
+        this.log('warn', 'RTCPeerConnection failed. Cleaning up session.', { state, conversationId: session.conversationId, sid: session.id, sessionType: session.sessionType });
+        this.onSessionTerminated(session, { condition: Constants.JingleReasonCondition.FailedTransport });
+      }
     });
 
     session.on('terminated', this.onSessionTerminated.bind(this, session));